Minor refactoring + boolean operators
lex now has it's own dir

CMake has been refactored slightly to group src files by subdir

argparse implementation has been shifted around to clear up main()
