+++ src/startup.cpp
@@ -1,6 +1,37 @@
-// This file builds the runtime
-
 #include <gearlang/func.hpp>
+#include <argparse/argparse.hpp>
+
+void init_program(argparse::ArgumentParser& program) {
+    program.add_description("Gearlang compiler");
+
+    program.add_argument("input")
+        .help("Input source file");
+
+    program.add_argument("-o", "--output")
+        .help("Output executable file")
+        .default_value(std::string("a.out"));
+
+    program.add_argument("-c", "--object")
+        .help("Emit object file")
+        .flag();
+
+    program.add_argument("-S", "--emit-llvm")
+        .help("Emit LLVM IR file")
+        .flag();
+
+    program.add_argument("-v", "--verbose")
+        .help("Enable verbose output")
+        .flag();
+    
+    program.add_argument("--dump-tokens")
+        .help("Print token output. Used for debugging the compiler")
+        .flag();
+
+    program.add_argument("--dump-ast")
+        .help("Print the AST output. Used for debugging the compiler")
+        .flag();
+
+}
 
 llvm::Function* build_runtime(Context& ctx) { 
     // Reworking this to be easier and just call main
